+++ src/ready.ts
@@ -1,0 +1,14 @@
+//#region Ready callback
+/**
+ * Execute a function when the page is ready (similar to jQuery ready function)
+ * @param fn Function. The function to execute
+ */
+export const ready = fn => {
+  if ((<any>document).attachEvent ? document.readyState === 'complete' : document.readyState !== 'loading') {
+    fn()
+  } else {
+    document.addEventListener('DOMContentLoaded', fn)
+  }
+}
+
+//#endregion Ready callback